requirements.txt
requests==2.31.0
pandas==2.1.4
streamlit==1.29.0
plotly==5.18.0
python-dotenv==1.0.0
beautifulsoup4==4.12.2
lxml==4.9.3  # XML parser for BeautifulSoup
folium>=0.15.1
streamlit-folium>=0.18.0
numpy>=1.26.4
matplotlib>=3.8.3
seaborn>=0.13.2
altair>=5.2.0
geopy>=2.4.1
scikit-learn>=1.4.1.post1 